+++ Runtime/BoolSet.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections
+{
+    [CreateAssetMenu(menuName = "Collections/Sets/Bool Set")]
+    public class BoolSet : RuntimeSet<bool> { }
+}