+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2022-12-29
+
+### Added
+
+- `BoolSet` collection
+- `BoolSetWatcher` component
+- `AllowDuplicates` property for `RuntimeSet` collection
+- `AddRange(IEnumerable<T>)` method for `RuntimeSet` collection
+- Icons for `RuntimeSet` collections in the Editor
+
 ## [0.2.2] - 2022-12-08
 
 ### Fixed